Task: Mirr changes

+++ src/main/java/com/lardi/exception/web/NotFoundException.java
@@ -6,15 +6,16 @@ import com.lardi.exception.AppException;
  * Created by vitalii.levash on 18.04.2016.
  */
 public class NotFoundException extends AppException {
-    public NotFoundException(final Long entityId) {
+    public NotFoundException(final int entityId) {
         super(String.format("could not find '%d'.", entityId));
     }
 
-    public NotFoundException(final Long entityId, final String entityName) {
+    public NotFoundException(final int entityId, final String entityName) {
         super(String.format("could not find %s by '%d'.", entityName, entityId));
     }
 
     public NotFoundException(final String msg) {
         super(msg);
     }
+
 }
